% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.metareg_fit}
\alias{print.metareg_fit}
\title{Print Method for metareg_fit}
\usage{
\method{print}{metareg_fit}(x, ...)
}
\arguments{
\item{x}{A \code{metareg_fit} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for metareg_fit
}
