% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summary.R
\name{print_metareg_summary}
\alias{print_metareg_summary}
\title{Print a metareg_summary}
\usage{
print_metareg_summary(x)
}
\arguments{
\item{x}{A \code{metareg_summary} object.}
}
\description{
Print a metareg_summary
}
\keyword{internal}
